requirements.txt
acres==0.5.0
aiofiles==25.1.0
aistudio-sdk==0.3.8
albucore==0.0.24
albumentations==2.0.8
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
asn1crypto==1.5.1
astor==0.8.1
asttokens==3.0.0
auto-py-to-exe==2.48.1
bce-python-sdk==0.9.68
beautifulsoup4==4.14.3
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2025.8.3
cffi==2.0.0
chardet==7.4.0.post2
charset-normalizer==3.4.3
ci-info==0.4.0
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
crc32c==2.8
cycler==0.12.1
Cython==3.2.4
debugpy==1.8.15
decorator==5.2.1
Django==5.2.4
Eel==0.18.2
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.0
fastapi==0.118.0
filelock==3.25.2
fire==0.7.1
fitz==0.0.1.dev2
fonttools==4.62.1
frontend==0.0.3
fsspec==2026.3.0
future==1.0.0
fuzzywuzzy==0.18.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.2.4
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.10
ImageIO==2.37.3
imgaug==0.4.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.5.0
lazy-loader==0.5
lmdb==2.1.1
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdurl==0.1.2
modelscope==1.35.3
nest-asyncio==1.6.0
networkx==3.6.1
nibabel==5.4.2
nipype==1.11.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==25.0
paddleocr==2.9.0
paddlepaddle==2.6.2
paddlex==3.4.2
pandas==2.3.1
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pefile==2023.2.7
pillow==12.1.1
platformdirs==4.3.8
prettytable==3.17.0
prompt_toolkit==3.0.51
protobuf==3.20.2
prov==2.1.1
psutil==7.0.0
pure_eval==0.2.3
puremagic==1.30
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMuPDF==1.24.14
pyodbc==5.2.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.2.1
python3-saml==1.16.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pyxnat==1.6.4
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.14.3
rdflib==7.6.0
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.19.1
scikit-image==0.26.0
scipy==1.17.1
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.48.0
stringzilla==4.6.0
termcolor==3.3.0
tifffile==2026.3.3
tornado==6.5.1
tqdm==4.67.3
traitlets==5.14.3
traits==7.1.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.12.0
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.13
whitenoise==6.9.0
xmlsec==1.3.17
zope.event==6.0
zope.interface==8.0.1
